Testing completato

+++ wannabet/src/main/java/is/project/wannabet/controller/AccountRegistratoController.java
@@ -1,9 +1,13 @@
 package is.project.wannabet.controller;
 
 import is.project.wannabet.model.AccountRegistrato;
+import is.project.wannabet.security.AuthenticationRequestAccountCheck;
 import is.project.wannabet.service.AccountRegistratoService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,6 +20,10 @@ public class AccountRegistratoController {
     @Autowired
     private AccountRegistratoService service;
 
+
+    @Autowired
+    private AuthenticationRequestAccountCheck accountCheck;
+
     @GetMapping
     @PreAuthorize("hasRole('ADMIN')")
     public List<AccountRegistrato> getAllAccounts() {
@@ -24,19 +32,33 @@ public class AccountRegistratoController {
 
     @GetMapping("/{id}")
     @PreAuthorize("hasAnyRole('UTENTE', 'ADMIN')")
-    public Optional<AccountRegistrato> getAccountById(@PathVariable Long id) {
-        return service.getAccountById(id);
+    public Optional<AccountRegistrato> getAccountById(@PathVariable Long id, Authentication authentication) {
+        ResponseEntity<?> response = accountCheck.checkAccount(id, authentication);
+
+        if (response.getStatusCode() == HttpStatus.OK ||
+                authentication.getAuthorities().stream()
+                        .anyMatch(authority -> authority.getAuthority().equals("ROLE_ADMIN"))
+        ) {
+            return service.getAccountById(id);
+        } else
+            return Optional.empty();
     }
 
     @PostMapping
-    @PreAuthorize("hasRole('UTENTE')")
     public AccountRegistrato createAccount(@RequestBody AccountRegistrato account) {
         return service.saveAccount(account);
     }
 
     @DeleteMapping("/{id}")
-    @PreAuthorize("#id == authentication.principal.idAccount and hasRole('UTENTE')")
-    public void deleteAccount(@PathVariable Long id) {
-        service.deleteAccount(id);
+    @PreAuthorize("hasAnyRole('UTENTE', 'ADMIN')")
+    public void deleteAccount(@PathVariable Long id, Authentication authentication) {
+        ResponseEntity<?> response = accountCheck.checkAccount(id, authentication);
+
+        if (response.getStatusCode() == HttpStatus.OK ||
+                authentication.getAuthorities().stream()
+                        .anyMatch(authority -> authority.getAuthority().equals("ROLE_ADMIN"))
+        ) {
+            service.deleteAccount(id);
+        }
     }
 }
